+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Show app icons in the settings navigation. It's pretty.
 * Make accessible errors at `Login` page. Refs STCOR-430.
+* Stop filtering "missing permission" warnings from `react-intl`. Refs STCOR-424.
 
 ## [5.0.0](https://github.com/folio-org/stripes-core/tree/v5.0.0) (2020-05-19)
 [Full Changelog](https://github.com/folio-org/stripes-core/compare/v4.1.0...v5.0.0)